ai-saas-week1/app/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic-settings==2.2.1
sqlalchemy==2.0.31
asyncpg==0.30.0
redis==5.0.1
python-multipart==0.0.9
greenlet==3.4.0
